backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sfap-backend"
version = "0.1.0"
description = "Multilingual Student Feedback Analysis Platform - Backend"
requires-python = ">=3.11,<3.12"
dependencies = [
    "pandas>=2.1",
    "numpy>=1.25",
    "scikit-learn>=1.3",
    "scipy>=1.11",
    "semopy>=2.3",
    "factor-analyzer>=0.5",
    "fastapi>=0.104",
    "uvicorn[standard]>=0.24",
    "pydantic>=2.5",
    "pyyaml>=6.0",
    "click>=8.1",
    "structlog>=23.2",
    "joblib>=1.3",
    "orjson>=3.9",
    "pingouin>=0.5",
    "httpx>=0.25",
    "aiofiles>=23.2",
    "python-multipart>=0.0.6",
    "pyarrow>=23.0.1",
]

[project.optional-dependencies]
transformers = [
    "transformers>=4.35",
    "torch>=2.1",
    "peft>=0.7",
]
fasttext = [
    "fasttext-wheel>=0.9",
]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "pytest-asyncio>=0.21",
    "httpx>=0.25",
]

[project.scripts]
sfap = "src.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["../tests/backend"]
pythonpath = ["."]
asyncio_mode = "auto"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
]

[dependency-groups]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "pytest-asyncio>=0.21",
    "httpx>=0.25",
]
